Fix snake case issues
A few operations and models are affected by a small problem where a
character gets isolated. This fixes it.
