+++ .generator/src/generator/templates/model_generic.j2
@@ -2,7 +2,7 @@
 {%- if refs %}
 def lazy_import():
 {%- for ref in refs %}
-    from {{ package }}.{{ version }}.model.{{ ref|snake_case }} import {{ ref }}
+    from {{ package }}.{{ version }}.model.{{ ref|model_name }} import {{ ref }}
 {%- endfor %}
 {# keep new line #}
 {%- for ref in refs %}
