+++ .generator/src/generator/templates/model_simple.j2
@@ -1,7 +1,7 @@
 {%- set ref = get_type_for_items(model) %}
 {%- if ref  %}
 def lazy_import():
-    from {{ package }}.{{ version }}.model.{{ ref|snake_case }} import {{ ref }}
+    from {{ package }}.{{ version }}.model.{{ ref|model_name }} import {{ ref }}
 
     globals()["{{ ref }}"] = {{ ref }}
 {%- endif %}
