+++ .generator/src/generator/templates/models.j2
@@ -1,4 +1,4 @@
 {%- for model in models %}
-from {{ package }}.{{ version }}.model.{{ model|snake_case }} import {{ model }}
+from {{ package }}.{{ version }}.model.{{ model|model_name }} import {{ model }}
 {%- endfor %}
