+++ src/datadog_api_client/v1/api/usage_metering_api.py
@@ -31,7 +31,7 @@ from datadog_api_client.v1.model.hourly_usage_attribution_usage_type import Hour
 from datadog_api_client.v1.model.usage_incident_management_response import UsageIncidentManagementResponse
 from datadog_api_client.v1.model.usage_indexed_spans_response import UsageIndexedSpansResponse
 from datadog_api_client.v1.model.usage_ingested_spans_response import UsageIngestedSpansResponse
-from datadog_api_client.v1.model.usage_io_t_response import UsageIoTResponse
+from datadog_api_client.v1.model.usage_iot_response import UsageIoTResponse
 from datadog_api_client.v1.model.usage_logs_response import UsageLogsResponse
 from datadog_api_client.v1.model.usage_logs_by_retention_response import UsageLogsByRetentionResponse
 from datadog_api_client.v1.model.usage_logs_by_index_response import UsageLogsByIndexResponse
