+++ src/datadog_api_client/v2/model/authn_mapping_attributes.py
