+++ src/datadog_api_client/v2/model/authn_mapping_update_attributes.py
