+++ src/datadog_api_client/v2/model/authn_mappings_type.py
