+++ src/datadog_api_client/v2/models/__init__.py
@@ -36,22 +36,22 @@ from datadog_api_client.v2.model.audit_logs_response_status import AuditLogsResp
 from datadog_api_client.v2.model.audit_logs_search_events_request import AuditLogsSearchEventsRequest
 from datadog_api_client.v2.model.audit_logs_sort import AuditLogsSort
 from datadog_api_client.v2.model.audit_logs_warning import AuditLogsWarning
-from datadog_api_client.v2.model.auth_n_mapping import AuthNMapping
-from datadog_api_client.v2.model.auth_n_mapping_attributes import AuthNMappingAttributes
-from datadog_api_client.v2.model.auth_n_mapping_create_attributes import AuthNMappingCreateAttributes
-from datadog_api_client.v2.model.auth_n_mapping_create_data import AuthNMappingCreateData
-from datadog_api_client.v2.model.auth_n_mapping_create_relationships import AuthNMappingCreateRelationships
-from datadog_api_client.v2.model.auth_n_mapping_create_request import AuthNMappingCreateRequest
-from datadog_api_client.v2.model.auth_n_mapping_included import AuthNMappingIncluded
-from datadog_api_client.v2.model.auth_n_mapping_relationships import AuthNMappingRelationships
-from datadog_api_client.v2.model.auth_n_mapping_response import AuthNMappingResponse
-from datadog_api_client.v2.model.auth_n_mapping_update_attributes import AuthNMappingUpdateAttributes
-from datadog_api_client.v2.model.auth_n_mapping_update_data import AuthNMappingUpdateData
-from datadog_api_client.v2.model.auth_n_mapping_update_relationships import AuthNMappingUpdateRelationships
-from datadog_api_client.v2.model.auth_n_mapping_update_request import AuthNMappingUpdateRequest
-from datadog_api_client.v2.model.auth_n_mappings_response import AuthNMappingsResponse
-from datadog_api_client.v2.model.auth_n_mappings_sort import AuthNMappingsSort
-from datadog_api_client.v2.model.auth_n_mappings_type import AuthNMappingsType
+from datadog_api_client.v2.model.authn_mapping import AuthNMapping
+from datadog_api_client.v2.model.authn_mapping_attributes import AuthNMappingAttributes
+from datadog_api_client.v2.model.authn_mapping_create_attributes import AuthNMappingCreateAttributes
+from datadog_api_client.v2.model.authn_mapping_create_data import AuthNMappingCreateData
+from datadog_api_client.v2.model.authn_mapping_create_relationships import AuthNMappingCreateRelationships
+from datadog_api_client.v2.model.authn_mapping_create_request import AuthNMappingCreateRequest
+from datadog_api_client.v2.model.authn_mapping_included import AuthNMappingIncluded
+from datadog_api_client.v2.model.authn_mapping_relationships import AuthNMappingRelationships
+from datadog_api_client.v2.model.authn_mapping_response import AuthNMappingResponse
+from datadog_api_client.v2.model.authn_mapping_update_attributes import AuthNMappingUpdateAttributes
+from datadog_api_client.v2.model.authn_mapping_update_data import AuthNMappingUpdateData
+from datadog_api_client.v2.model.authn_mapping_update_relationships import AuthNMappingUpdateRelationships
+from datadog_api_client.v2.model.authn_mapping_update_request import AuthNMappingUpdateRequest
+from datadog_api_client.v2.model.authn_mappings_response import AuthNMappingsResponse
+from datadog_api_client.v2.model.authn_mappings_sort import AuthNMappingsSort
+from datadog_api_client.v2.model.authn_mappings_type import AuthNMappingsType
 from datadog_api_client.v2.model.cloud_workload_security_agent_rule_attributes import (
     CloudWorkloadSecurityAgentRuleAttributes,
 )
